edit folder name
