--- Stock_ENV/gym_stock.egg-info/PKG-INFO
@@ -1,11 +1,0 @@
-Metadata-Version: 1.0
-Name: gym-stock
-Version: 0.0.1
-Summary: UNKNOWN
-Home-page: UNKNOWN
-Author: UNKNOWN
-Author-email: UNKNOWN
-License: UNKNOWN
-Description-Content-Type: UNKNOWN
-Description: UNKNOWN
-Platform: UNKNOWN
