--- Stock_ENV/gym_stock/envs/test_stock_env.py
@@ -1,6 +1,0 @@
-import gym
-import gym_stock
-
-
-env = gym.make('stock-v0')
-
