+++ gym_stock/gym_stock.egg-info/requires.txt
