+++ gym_stock/gym_stock.egg-info/top_level.txt
