+++ gym_stock/gym_stock/__init__.py
@@ -1,0 +1,6 @@
+from gym.envs.registration import register
+
+register(
+    id='stock-v0',
+    entry_point='gym_stock.envs:StockEnv',
+)
