+++ gym_stock/gym_stock/envs/test_stock_env.py
@@ -1,0 +1,6 @@
+import gym
+import gym_stock
+
+
+env = gym.make('stock-v0')
+
